Create COVID_Cases for Indian States-Maps in R
Hello
I am quite a novice when it comes to R and R programming. Having been just introduced to the same, I was trying to use up a lot of hacks and reviews of other similar projects to try and build a r Script that will help me Map Covid19 cases of Different Indians states and visualize the same as Choropleth District Maps for states like Assam. 
To a very basic extend I was able to Create what I was looking for . However, I have been unable to Type the names of the Districts and the number of Cases in the Map Visualization.  I tried "geom_text(data=name_lat_lon, aes(long, lat, label = NAME_1), size=2)" However, the same is returning an error in fortify(data):object 'name_lat_long' not found. I am really looking for an improvement in this above code.